ClipOnly2: remove long double precision

+++ src/ClipOnly2/ClipOnly2.c
@@ -19,12 +19,12 @@ typedef struct {
 	const float* input[2];
 	float* output[2];
 
-	long double lastSampleL;
-	long double intermediateL[16];
+	double lastSampleL;
+	double intermediateL[16];
 	bool wasPosClipL;
 	bool wasNegClipL;
-	long double lastSampleR;
-	long double intermediateR[16];
+	double lastSampleR;
+	double intermediateR[16];
 	bool wasPosClipR;
 	bool wasNegClipR; //Stereo ClipOnly2
 } ClipOnly2;
